Fix access issue with shared memory
The memory mapped buffer is written from kernel space. Hence, the
user space read-write mutex cannot be locked just prior to write.
However, with a single kernel space writer, and multiple user space
readers, a mutex is required.

To resolve this impasse, a user-space buffer is used. Each received
frame is copied to it with mutex protection. Readers read from this
buffer with mutex protection.